+++ src/pyne_runtime/request/__init__.py
@@ -6,16 +6,34 @@ from .errors import PyneInvalidSymbolError, PyneRequestError
 from .eval import RequestEvalContext, RequestValues
 from .lower_tf import LowerTimeframeSeries
 from .module import RequestModule
-from .provider import DataProvider
+from .provider import (
+    DataProvider,
+    OHLCVBar,
+    RequestCapabilities,
+    RequestCapabilityProvider,
+    RequestMetadata,
+    RequestMetadataProvider,
+    RequestSessionMetadata,
+    RequestSymbolMetadata,
+    RequestTimeframeMetadata,
+)
 
 __all__ = [
     "BarMergeNamespace",
     "DataProvider",
     "LowerTimeframeSeries",
+    "OHLCVBar",
     "PyneInvalidSymbolError",
     "PyneRequestError",
+    "RequestCapabilities",
+    "RequestCapabilityProvider",
     "RequestEvalContext",
+    "RequestMetadata",
+    "RequestMetadataProvider",
     "RequestModule",
+    "RequestSessionMetadata",
+    "RequestSymbolMetadata",
+    "RequestTimeframeMetadata",
     "RequestValues",
     "barmerge",
 ]